fix(T09): magic numbers / divide todos constants on groups
