+++ src/common/constants/TodoConstants.ts
@@ -1,66 +1,67 @@
-export enum TodoManagementActions {
-  ADD_TODO = "ADD_TODO",
-  DELETE_TODO = "DELETE_TODO",
-  DELETE_COMPLETED_TODOS = "DELETE_COMPLETED_TODOS",
-}
-
-export enum TodoEditingActions {
-  EDIT_TODO = "EDIT_TODO",
-  SET_TODO_TITLE = "SET_TODO_TITLE",
-  SET_COMPLETED_TODO = "SET_COMPLETED_TODO",
-  SET_CURRENT_TODO = "SET_CURRENT_TODO",
-}
-
-export enum TodoFilteringActions {
-  SET_FILTRATION_VALUE = "SET_FILTRATION_VALUE",
-  SEARCH_TODO = "SEARCH_TODO",
-  SEARCH_VALUE = "SEARCH_VALUE",
-}
-
-export const TodoConstants = {
-  ...TodoManagementActions,
-  ...TodoEditingActions,
-  ...TodoFilteringActions,
-};
-
-export const enum TodoValidateFields {
-  TODO_TITLE = "todoTitle",
-  CREATED_DATE = "createdDate",
-  EXPIRATION_DATE = "expirationDate",
-}
-
-export const enum TodoNotificationMessages {
-  MAX_LENGTH = "The maximum length of the title mustn't exceed 120 characters",
-  EMPTY_TITLE = "You can't to create empty todo",
-  REQUIRED_FIELD = "Required field",
-  DELETE_COMPLETED_TODOS = "Successfully cleared the completed tasks",
-  DELETE_TODO = "Successfully deleted the todo",
-}
-
-export const enum TodoConfirmMessages {
-  DELETE_TODO = "delete current todo",
-  DELETE_COMPLETED_TODOS = "delete completed todos",
-}
-
-export const enum TodoTimeConstants {
-  ADDITION_TODO_TIME = 5,
-  RESET_ADDITION_TODO_TIME = -5,
-  TIME_INTERVAL = 1,
-}
-
-export const TodoCurrentFilter = {
-  ALL: "All",
-  ACTIVE: "Active",
-  COMPLETED: "Completed",
-};
-
-export const TodoCurrentFilterArray = Object.values(TodoCurrentFilter);
-
-export const enum DateInfo {
-  HOURS_IN_DAY = 24,
-}
-
-export const TodoValidateData = {
-  MAX_TITLE_LENGTH: 120,
-  MIN_TITLE_LENGTH: 1,
-};
+// export enum TodoManagementActions {
+//   ADD_TODO = "ADD_TODO",
+//   DELETE_TODO = "DELETE_TODO",
+//   DELETE_COMPLETED_TODOS = "DELETE_COMPLETED_TODOS",
+// }
+//
+// export enum TodoEditingActions {
+//   EDIT_TODO = "EDIT_TODO",
+//   SET_TODO_TITLE = "SET_TODO_TITLE",
+//   SET_COMPLETED_TODO = "SET_COMPLETED_TODO",
+//   SET_CURRENT_TODO = "SET_CURRENT_TODO",
+// }
+//
+// export enum TodoFilteringActions {
+//   SET_FILTRATION_VALUE = "SET_FILTRATION_VALUE",
+//   SEARCH_TODO = "SEARCH_TODO",
+//   SEARCH_VALUE = "SEARCH_VALUE",
+// }
+//
+// export const TodoConstants = {
+//   ...TodoManagementActions,
+//   ...TodoEditingActions,
+//   ...TodoFilteringActions,
+// };
+//
+// export const enum TodoValidateFields {
+//   TODO_TITLE = "todoTitle",
+//   CREATED_DATE = "createdDate",
+//   EXPIRATION_DATE = "expirationDate",
+// }
+//
+// export const enum TodoNotificationMessages {
+//   MAX_LENGTH = "The maximum length of the title mustn't exceed 120 characters",
+//   EMPTY_TITLE = "You can't to create empty todo",
+//   REQUIRED_FIELD = "Required field",
+//   DELETE_COMPLETED_TODOS = "Successfully cleared the completed tasks",
+//   DELETE_TODO = "Successfully deleted the todo",
+// }
+//
+// export const enum TodoConfirmMessages {
+//   DELETE_TODO = "delete current todo",
+//   DELETE_COMPLETED_TODOS = "delete completed todos",
+// }
+//
+// export const enum TodoTimeConstants {
+//   ADDITION_TODO_TIME = 5,
+//   RESET_ADDITION_TODO_TIME = -5,
+//   TIME_INTERVAL = 1,
+// }
+//
+// export const TodoCurrentFilter = {
+//   ALL: "All",
+//   ACTIVE: "Active",
+//   COMPLETED: "Completed",
+// };
+//
+// export const TodoCurrentFilterArray = Object.values(TodoCurrentFilter);
+//
+// export const enum DateInfo {
+//   HOURS_IN_DAY = 24,
+// }
+//
+// export const TodoValidateData = {
+//   MAX_TITLE_LENGTH: 120,
+//   MIN_TITLE_LENGTH: 1,
+//   MAX_DATE_LENGTH: 16,
+// };
