+++ src/common/constants/TodoConstants/TodoFilters.ts
@@ -1,0 +1,7 @@
+export const TodoCurrentFilter = {
+  ALL: "All",
+  ACTIVE: "Active",
+  COMPLETED: "Completed",
+};
+
+export const TodoCurrentFilterArray = Object.values(TodoCurrentFilter);
