+++ src/common/constants/TodoConstants/TodoTimeConstants.ts
@@ -1,0 +1,9 @@
+export const enum TodoTimeConstants {
+  ADDITION_TODO_TIME = 5,
+  RESET_ADDITION_TODO_TIME = -5,
+  TIME_INTERVAL = 1,
+}
+
+export const enum TodoDateInfo {
+  HOURS_IN_DAY = 24,
+}
