+++ src/common/constants/TodoConstants/TodoValidationData.ts
@@ -1,0 +1,5 @@
+export const TodoValidateData = {
+  MAX_TITLE_LENGTH: 120,
+  MIN_TITLE_LENGTH: 1,
+  MAX_DATE_LENGTH: 16,
+};
