+++ src/utils/setSelectedDate.ts
@@ -1,4 +1,5 @@
-import { TodoTimeConstants, TodoValidateFields } from "@/common/constants/TodoConstants.ts";
+import { TodoTimeConstants } from "@/common/constants/TodoConstants/TodoTimeConstants.ts";
+import { TodoValidateFields } from "@/common/constants/TodoConstants/TodoValidation.ts";
 import { setMinutesToDate } from "@/utils/setTimeToDate.ts";
 import { Dispatch, SetStateAction } from "react";
